+++ src/struct/collection.h
@@ -15,7 +15,7 @@
 // back to it — one subtract, nothing more.
 
 typedef struct Collection {
-    uint32_t typeId;        // mirror of the block-header type (for debug)
+    uint64_t typeId;        // mirror of the block-header type (for debug)
     uint32_t activeCount;   // number of live elements
     uint32_t elementClass;  // class of elements (Map: key class)
     uint32_t stride;        // bytes per element (Map: val class)
@@ -24,7 +24,7 @@ typedef struct Collection {
     uint8_t *data;          // element / slot buffer
 } Collection;
 
-uint32_t Collection_type(Collection *c);
+uint64_t Collection_type(Collection *c);
 uint32_t Collection_size(Collection *c);
 uint32_t Collection_length(Collection *c);
 bool Collection_isEmpty(Collection *c);
